+++ mcxr-play/src/main/java/net/sorenon/mcxr/play/gui/QuickMenu.java
@@ -1,4 +1,4 @@
-package net.sorenon.mcxr.play.mixin.gui;
+package net.sorenon.mcxr.play.gui;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 import net.fabricmc.loader.api.FabricLoader;
@@ -22,7 +22,7 @@ public class QuickMenu extends Screen {
         ArrayList<Button> QuickMenuButtons = new ArrayList<>();
 
         QuickMenuButtons.add(new Button((this.width/2) - 25, this.height/2, 70, 20, new TranslatableComponent("QuickChat"), (button ) -> {
-            Minecraft.getInstance().setScreen(new net.sorenon.mcxr.play.mixin.gui.QuickChat("QuickChat"));
+            Minecraft.getInstance().setScreen(new net.sorenon.mcxr.play.gui.QuickChat("QuickChat"));
         }));
         if (FabricLoader.getInstance().isModLoaded("voicechat")) {
             SimpleVoiceChatCompat.createButton(QuickMenuButtons, this.width, this.height);
