+++ mcxr-play/src/main/java/net/sorenon/mcxr/play/mixin/EditBoxMixin.java
@@ -4,7 +4,7 @@ import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.components.EditBox;
 import net.minecraft.client.gui.screens.ChatScreen;
 import net.minecraft.network.chat.TranslatableComponent;
-import net.sorenon.mcxr.play.mixin.gui.Vr2DKeyboard;
+import net.sorenon.mcxr.play.gui.Vr2DKeyboard;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
